Fix serial port capabilities in setup.py depending on distro
The group granting the serial port capabilities is not standard and depends on
the used distribution. To not break the complete devices_install chain because
of a group mismatch, check for the group before using usermod. If the group is
not found, a self-explanatory exception is raised instead of the "An error
occured during Serial port capabilities installation" message.