Improve UART-based WHAD interface disconnection detection.
When a UART-based WHAD interface returns an empty string while a call to `select()`
indicates the device has some new data to be read, that means this device has an issue
(disconnected, malfunction, ...) and must be considered as "not ready".

We now raise a `WhadDeviceNotReady` exception when such situation occurs, allowing
the framework to catch it and handle it nicely in command-line tools.