Added a new feature in `WhadDeviceConnector` to lock/unlock a connector, keeping messages in a parallel waiting queue when locked.
This feature allows tools to create a WHAD connector that may catch incoming messages and to delay their processing until some
conditions are met. This is actually used in `wble-connect` when establishing a BLE connection to a device to let a chained tool
process any incoming message that may be sent by the target device before the chained tool connects to `wble-connect`. Once the
chained tool connected, the connector is unlocked and all the pending messages are then processed as they were just received,
avoiding any race condition or timing issue.

This feature greatly improves Man-in-the-Middle stability, partly solving issue #52.