Added 802.15.4 packet dissection watchdogs in WHAD hub to avoid unwanted parsing exceptions raised by Scapy.
These exceptions raised by Scapy broke the workflow of wzb-enddevice and caused a lot of issues due to
malformatted 802.15.4 packets, scapy trying deseperately to interpret payload as ZigBee or 6LoWPAN content.

Not sure this solution will fit non-standard protocols based on IEEE 802.15.4 MAC, but for now it will
help a lot avoiding unwanted crashes in Zigbee and RF4CE tools and connectors.