Improved Central and Peripheral to handle events using queues rather than callbacks.
Previous code was using callbacks that messed up with WHAD device message processing,
as we were tempted to start an action on a specific event and therefore send more
messages while blocking the incoming messages processing loop. Switching to a
dedicated queue for each Central and Peripheral, running in a dedicated thread,
allows to continue message processing while the event manager thread catches any
event and dispatches to Central or Peripheral class.

It is a poor-mans async communication mechanism that would be largely improved in
an upcoming rework of the asynchronous layer of WHAD (namely WhadDevice, WhadDeviceConnector
and their related background tasks). But that's another story.