Fixed a bug in the way indication callbacks were called.
On registration for notification for a specific characteristic,
the callback provided by the user is wrapped inside a closure that
injects the related characteristic object as first argument, replacing
the characteristic handle sent by our GATT stack. This way, the user
can directly get a reference on the characteristic, read its value, etc.

This was not the case for indications, so I added the same closure to
inject the characteristic object as we do for notifications, in order
to keep callbacks consistent.