Rolled back to `whad.device` namespace instead of `whad.hw`, renamed `Interface` and `VirtualInterface` to `Device` and `VirtualDevice`.
Following a discussion with @RCayre, we figured out that moving to a completely different namespace (while keeping the old one for a
smooth transition) will mess up with our documentation and more importantly will confuse users as it differs from the original naming.
As the decision to move `whad.device` to `whad.hw` was motivated by a will of clarity, the consequences of such a renaming would be
in total contradiction with the intended goal.

Therefore, the following changes have been decided and implemented:
- previous `whad.hw.Interface` class has been moved to `whad.device.Device` with an alias of `whad.device.WhadDevice`
- previous `whad.hw.VirtualInterface` class has been moved to `whad.device.VirtualDevice` with an alias of `whad.device.WhadVirtualDevice`
- `WhadDevice` and `WhadConnector` have been renamed to `Device` and `Connector`, aliases for old names have been created to maintain compatibility
- connectors have been reworked to use these new classes