Added a new stack model feature called 'flavor', that allows creation of flavors of a specific stack
This flavor mechanism provides a way to instantiate a stack with a specific flavor in which some
layers's implementation may differ from the one defined by default. It is also a neat way for
advanced user to replace a specific layer with their in their own flavor and instantiate this
flavored stack without impacting the default implementation.