Modified peripheral's `wait_connection()` and `wait_disconnection()` methods to use `Event` objects.
Removing calls to `time.sleep()` is always a good idea as it is best to leave the OS handle waiting
jobs through mutexes rather than relying on `sleep()` ...

`threading`'s Event objects are perfect for this, as `wait_connection()` and `wait_disconnection()`
methods are basically called from the application's main thread. This allows Peripheral's callbacks
to notify the main thread when a central connects to or disconnects from a peripheral.