Changed expected minimal ButteRFly version (was 1.1.0, now 1.0.2 due to our release process)
ButteRFly version number has mistakenly been moved 1.0.1 (previous release version number)
to 1.0.2 during the last release, overwriting the previously modified version number that
was set to 1.1.0.

The problem is it causes WHAD to display a message to the user asking them to upgrade to
the latest version while they are using in fact the latest one :S.

This fix solves this by checking the correct ButteRFly version (1.0.2).